// ExistingCode/src/data-access/domain/tables/drillplan/schema.ts
export const DrillPlanTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	drillPlanId: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	organization: z.string().max(30),
	/**
	 * dataType: nvarchar
	 */
	drillPattern: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	project: z.string().max(100),
	/**
	 * dataType: nvarchar
	 */
	prospect: z.string().max(30).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	target: z.string().max(100).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	subTarget: z.string().max(100).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	pit: z.string().max(30).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	phase: z.string().max(30).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	zone: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	tenement: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	holeType: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	holePurpose: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	holePurposeDetail: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	drillType: z.string().max(50).nullable().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	drillPriority: z.number().int().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	oDSPriority: z.number().int().optional(),
	/**
	 * dataType: smallint
	 * defaultValue: ((0))
	 */
	priority: z.number().int().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	grid: z.string().max(50).nullable().optional(),
	/**
	 * dataType: decimal
	 */
	plannedEasting: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	plannedNorthing: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	plannedRL: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	plannedDip: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	plannedAzimuth: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	plannedTotalDepth: z.number().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	plannedStartDt: z.any().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	plannedCompleteDt: z.any().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	waterTableDepth: z.number().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	sitePrep: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	infillTarget: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	tWF: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	plannedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	qCInsertionRuleId: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	dataSource: z.string().max(255),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	reportIncludeInd: z.any().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	validationStatus: z.number().int().optional(),
	/**
	 * dataType: nvarchar
	 */
	validationErrors: z.string().max(4000).nullable().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	rowStatus: z.number().int().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	supersededById: z.any().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((1))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (CONVERT([nvarchar](50),coalesce(session_context(N'UserName'),suser_sname())))
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});